+++ app/models/user_survey.rb
@@ -2,8 +2,9 @@
 
 # Tracks survey completion.
 class UserSurvey < ApplicationRecord
+  # Validations
+  validates :user_id, uniqueness: { scope: [:event, :design] }
+
   # Relationships
   belongs_to :user
-
-  # Methods
 end