+++ db/migrate/20180205211149_add_unique_index_to_brain_tests.rb
@@ -1,0 +1,5 @@
+class AddUniqueIndexToBrainTests < ActiveRecord::Migration[5.2]
+  def change
+    add_index :brain_tests, [:user_id, :event, :battery_number, :test_number], unique: true, name: "index_user_brain_tests"
+  end
+end