+++ db/migrate/20180205211209_add_unique_index_to_user_surveys.rb
@@ -1,0 +1,5 @@
+class AddUniqueIndexToUserSurveys < ActiveRecord::Migration[5.2]
+  def change
+    add_index :user_surveys, [:user_id, :event, :design], unique: true, name: "index_user_slice_surveys"
+  end
+end